+++ final_task/src/main/java/domain/Order.java
@@ -1,5 +1,6 @@
 package domain;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 import java.util.List;
 
@@ -8,7 +9,7 @@ public class Order extends Entity {
 
     private  TypePay type_pay;
     private  Date date;
-    private  double total_price;
+    private BigDecimal total_price;
 
     private  User user;
     private  TypeDelivery delivery;
@@ -37,11 +38,11 @@ public class Order extends Entity {
         this.date = date;
     }
 
-    public double getTotal_price() {
+    public BigDecimal getTotal_price() {
         return total_price;
     }
 
-    public void setTotal_price(double total_price) {
+    public void setTotal_price(BigDecimal total_price) {
         this.total_price = total_price;
     }
